+++ server/storage_index_test.go
@@ -139,13 +139,13 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 			t.Fatal(err.Error())
 		}
 
-		entries, err := storageIdx.List(ctx, uuid.Nil, indexName1, "", maxEntries1, []string{}) // Match all
+		entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName1, "", maxEntries1, []string{}, "") // Match all
 		if err != nil {
 			t.Fatal(err.Error())
 		}
 		assert.Len(t, entries.Objects, 2, "indexed results length was not 2")
 
-		entries, err = storageIdx.List(ctx, uuid.Nil, indexName2, "", maxEntries1, []string{}) // Match all
+		entries, _, err = storageIdx.List(ctx, uuid.Nil, indexName2, "", maxEntries1, []string{}, "") // Match all
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -186,7 +186,7 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 			t.Fatal(err.Error())
 		}
 
-		entries, err := storageIdx.List(ctx, uuid.Nil, indexName1, "+value.three:3", maxEntries1, []string{})
+		entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName1, "+value.three:3", maxEntries1, []string{}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -289,7 +289,7 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 			t.Fatal(err.Error())
 		}
 
-		entries, err := storageIdx.List(ctx, uuid.Nil, indexName2, "", maxEntries2, []string{})
+		entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName2, "", maxEntries2, []string{}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -316,6 +316,114 @@ func TestLocalStorageIndex_Write(t *testing.T) {
 }
 
 func TestLocalStorageIndex_List(t *testing.T) {
+	t.Run("paginates correctly", func(t *testing.T) {
+		db := NewDB(t)
+		defer db.Close()
+
+		ctx := context.Background()
+
+		nilUid := uuid.Nil
+
+		u1 := uuid.Must(uuid.NewV4())
+		InsertUser(t, db, u1)
+
+		indexName := "test_index_only"
+		collection := "test_collection"
+		key := "key"
+		maxEntries := 10
+
+		valueOneBytes, _ := json.Marshal(map[string]any{
+			"one": 1,
+		})
+		valueOne := string(valueOneBytes)
+		valueTwoBytes, _ := json.Marshal(map[string]any{
+			"two": 2,
+		})
+		valueTwo := string(valueTwoBytes)
+		valueThreeBytes, _ := json.Marshal(map[string]any{
+			"three": 3,
+		})
+		valueThree := string(valueThreeBytes)
+
+		storageIdx, err := NewLocalStorageIndex(logger, db, &StorageConfig{}, metrics)
+		if err != nil {
+			t.Fatal(err.Error())
+		}
+
+		if err := storageIdx.CreateIndex(ctx, indexName, collection, key, []string{"one", "two", "three"}, []string{}, maxEntries, true); err != nil {
+			t.Fatal(err.Error())
+		}
+
+		so1 := &StorageOpWrite{
+			OwnerID: nilUid.String(),
+			Object: &api.WriteStorageObject{
+				Collection: collection,
+				Key:        key,
+				Value:      valueOne,
+			},
+		}
+		so2 := &StorageOpWrite{
+			OwnerID: u1.String(),
+			Object: &api.WriteStorageObject{
+				Collection: collection,
+				Key:        key,
+				Value:      valueTwo,
+			},
+		}
+		so3 := &StorageOpWrite{
+			OwnerID: u1.String(),
+			Object: &api.WriteStorageObject{
+				Collection: collection,
+				Key:        key,
+				Value:      valueThree,
+			},
+		}
+
+		writeOps := StorageOpWrites{so1, so2, so3}
+
+		if _, _, err := StorageWriteObjects(context.Background(), logger, db, metrics, storageIdx, true, writeOps); err != nil {
+			t.Fatal(err.Error())
+		}
+
+		// Page 1
+		entries, cursor, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.two:2 value.three:3", 1, []string{}, "")
+		if err != nil {
+			t.Fatal(err.Error())
+		}
+		assert.Len(t, entries.Objects, 1, "indexed results did not match query params")
+		assert.NotEmptyf(t, cursor, "cursor was empty when there's more results")
+
+		// Page 2
+		entries, cursor, err = storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.two:2 value.three:3", 1, []string{}, cursor)
+		if err != nil {
+			t.Fatal(err.Error())
+		}
+		assert.Len(t, entries.Objects, 1, "indexed results did not match query params")
+		assert.NotEmptyf(t, cursor, "cursor was empty when there's more results")
+
+		// Page 3
+		entries, cursor, err = storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.two:2 value.three:3", 1, []string{}, cursor)
+		if err != nil {
+			t.Fatal(err.Error())
+		}
+		assert.Len(t, entries.Objects, 1, "indexed results did not match query params")
+		assert.Empty(t, cursor, "cursor was not empty on last page")
+
+		delOps := make(StorageOpDeletes, 0, len(writeOps))
+		for _, op := range writeOps {
+			delOps = append(delOps, &StorageOpDelete{
+				OwnerID: op.OwnerID,
+				ObjectID: &api.DeleteStorageObjectId{
+					Collection: op.Object.Collection,
+					Key:        op.Object.Key,
+				},
+			})
+		}
+		if _, err = StorageDeleteObjects(ctx, logger, db, storageIdx, true, delOps); err != nil {
+			t.Fatalf("Failed to teardown: %s", err.Error())
+		}
+	})
+
 	t.Run("when indexOnly is false, returns all matching results for query from the db", func(t *testing.T) {
 		db := NewDB(t)
 		defer db.Close()
@@ -385,7 +493,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 			t.Fatal(err.Error())
 		}
 
-		entries, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{})
+		entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -481,7 +589,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 			t.Fatal(err.Error())
 		}
 
-		entries, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{})
+		entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -490,7 +598,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 		assert.Equal(t, valueOne, entries.Objects[0].Value, "expected value retrieved from db did not match")
 		assert.Equal(t, valueThree, entries.Objects[1].Value, "expected value retrieved from db did not match")
 
-		sortEntries, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{"value.sort"})
+		sortEntries, _, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{"value.sort"}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -498,7 +606,7 @@ func TestLocalStorageIndex_List(t *testing.T) {
 		assert.Equal(t, valueOne, sortEntries.Objects[0].Value, "expected value retrieved from db did not match")
 		assert.Equal(t, valueThree, sortEntries.Objects[1].Value, "expected value retrieved from db did not match")
 
-		sortDescEntries, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{"-value.sort"})
+		sortDescEntries, _, err := storageIdx.List(ctx, uuid.Nil, indexName, "value.one:1 value.three:3", 10, []string{"-value.sort"}, "")
 		if err != nil {
 			t.Fatal(err.Error())
 		}
@@ -573,7 +681,7 @@ func TestLocalStorageIndex_Delete(t *testing.T) {
 		t.Fatal(err.Error())
 	}
 
-	entries, err := storageIdx.List(ctx, uuid.Nil, indexName, "", 10, []string{})
+	entries, _, err := storageIdx.List(ctx, uuid.Nil, indexName, "", 10, []string{}, "")
 	if err != nil {
 		t.Fatal(err.Error())
 	}
@@ -590,7 +698,7 @@ func TestLocalStorageIndex_Delete(t *testing.T) {
 		t.Fatal(err.Error())
 	}
 
-	entries, err = storageIdx.List(ctx, uuid.Nil, indexName, "", 10, []string{})
+	entries, _, err = storageIdx.List(ctx, uuid.Nil, indexName, "", 10, []string{}, "")
 	if err != nil {
 		t.Fatal(err.Error())
 	}